removed template
