--- npm/@libgs/template/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@libgs/",
-  "version": "1.0",
-  "os": "",
-  "arch": ""
-}
